+++ src/MockIO/Process/OutputDefault.hs
@@ -5,6 +5,10 @@ module MockIO.Process.OutputDefault
   ( OutputDefault( outDef ) )
 where
 
+-- bytestring --------------------------
+
+import qualified  Data.ByteString
+
 -- monadio-plus ------------------------
 
 import MonadIO.Process.ExitStatus( ExitStatus( ExitVal ) )
@@ -14,6 +18,10 @@ import MonadIO.Process.ExitStatus( ExitStatus( ExitVal ) )
 import Data.MoreUnicode.Text  ( 𝕋 )
 
 --------------------------------------------------------------------------------
+
+type 𝔹𝕊 = Data.ByteString.ByteString
+
+------------------------------------------------------------
 
 {- | Default values for mocked process output; exit 0 for command, empty texts
      where used. -}
@@ -29,25 +37,13 @@ instance OutputDefault () where
 instance OutputDefault 𝕋 where
   outDef = ""
 
-instance OutputDefault (𝕋,()) where
-  outDef = ("",())
+instance OutputDefault [𝕋] where
+  outDef = []
 
-instance OutputDefault ((),𝕋) where
-  outDef = ((),"")
+instance OutputDefault 𝔹𝕊 where
+  outDef = ""
 
-instance OutputDefault ((),()) where
-  outDef = ((),())
-
-instance OutputDefault (𝕋,𝕋) where
-  outDef = ("","")
-
-instance OutputDefault ([𝕋]) where
-  outDef = ([])
-
-instance OutputDefault ([𝕋],[𝕋]) where
-  outDef = ([],[])
-
-instance OutputDefault ξ ⇒ OutputDefault (ExitStatus, ξ) where
+instance (OutputDefault ξ, OutputDefault ξ') ⇒ OutputDefault (ξ, ξ') where
   outDef = (outDef, outDef)
 
 -- that's all, folks! ----------------------------------------------------------