+++ src/MockIO/T/Process.hs
@@ -11,6 +11,15 @@ import Data.Function  ( flip )
 import Data.Maybe     ( isJust )
 import GHC.Exts       ( IsList( toList ) )
 
+-- bytestring --------------------------
+
+import qualified  Data.ByteString  as  BS
+import Data.ByteString  ( ByteString )
+
+-- containers --------------------------
+
+import qualified Data.Set  as  Set
+
 -- containers-plus ---------------------
 
 import ContainersPlus.Insert  ( (⨭) )
@@ -39,7 +48,7 @@ import Log.LogEntry  ( LogEntry, logdoc )
 -- logging-effect ----------------------
 
 import Control.Monad.Log  ( MonadLog, Severity( Notice )
-                          , runPureLoggingT )
+                          , discardLogging, runPureLoggingT )
 
 -- mockio ------------------------------
 
@@ -65,6 +74,10 @@ import MonadIO.Process.CmdSpec        ( CmdArgs( CmdArgs ), CmdExe( CmdExe )
                                       , cmdArgs, cmdExe, expExitVal, mkCmd )
 import MonadIO.Process.ExitInfo       ( ExitInfo )
 import MonadIO.Process.ExitStatus     ( ExitStatus( ExitVal ), exitVal )
+import MonadIO.Process.MakeProc       ( MakeProc )
+import MonadIO.Process.OutputHandles  ( OutputHandles )
+import MonadIO.Process.ToMaybeTexts   ( ToMaybeTexts )
+import MonadIO.Temp                   ( testsWithTempfile )
 
 -- mtl ---------------------------------
 
@@ -77,7 +90,7 @@ import Prettyprinter.Render.Text  ( renderStrict )
 
 -- tasty-hunit -------------------------
 
-import Test.Tasty.HUnit  ( Assertion, assertBool, assertEqual )
+import Test.Tasty.HUnit  ( Assertion, assertBool, assertEqual, assertFailure )
 
 -- tasty-plus --------------------------
 
@@ -85,7 +98,7 @@ import TastyPlus  ( (≟), assertIOError, assertJust )
 
 -- text --------------------------------
 
-import Data.Text  ( isInfixOf, unlines, unpack )
+import Data.Text  ( isInfixOf, isSuffixOf, unlines, unpack )
 
 -- text-icu ----------------------------
 
@@ -98,17 +111,90 @@ import Data.Text.ICU  ( Regex, find, regex )
 
 import qualified  MockIOPlus.Paths  as  Paths
 
-import MockIO.Process                ( ꙩ, system )
+import MockIO.Process                ( ꙩ, sysN, system )
 import MockIO.Process.CmdRW          ( CmdRW( CmdR ) )
+import MockIO.Process.OutputDefault  ( OutputDefault )
+import MockIO.Process.MLCmdSpec      ( MLCmdSpec )
 import MockIO.Process.MLMakeIStream  ( MLMakeIStream )
 
 --------------------------------------------------------------------------------
 
-foo ∷ 𝕋
-foo = unlines [ "jimmy 7"
-              , "martyn 12"
-              , "marbyns 3"
-              ]
+type 𝔹𝕊 = ByteString
+
+{-| Grep a file using sysN, using different types of output, to test that sysN
+    and the differing output types each work.
+
+    We do this by running a simple grep against a temporary file, each time
+    checking that the result is as expected (in a given type).
+-}
+
+xx ∷ MLCmdSpec ξ → MLCmdSpec ξ
+xx x = x & expExitVal ⊢ Set.fromList [1]
+
+sysTests ∷ TestTree
+sysTests = testGroup "sysTests" $
+  let
+    foo ∷ 𝕋
+    foo = unlines [ "jimmy 7", "martyn 12", "marbyns 3" ]
+    check ∷ (OutputDefault γ, ToMaybeTexts γ, Printable β,
+             OutputHandles ζ γ, MakeProc ζ) ⇒
+            α → 𝕋 → ((ExitInfo,γ) → Assertion) → (α, β → IO ())
+
+    -- simplifed sysN, with a ProcError error, takes specifically an AbsFile,
+    -- 𝕋 args, an MLCmdSpec adjuster, discards logging and never mocks.
+    sysN' ∷ ∀ μ ξ ζ .
+            (MonadIO μ, OutputDefault ξ, ToMaybeTexts ξ, OutputHandles ζ ξ,
+             MakeProc ζ, MonadError ProcError μ) ⇒
+            AbsFile → [𝕋] → (MLCmdSpec ξ → MLCmdSpec ξ) → μ (ExitInfo, ξ)
+    sysN' p as f =
+      discardLogging ∘ flip runReaderT NoMock $ sysN (p, as, f)
+
+--    grp t f = sysN @ProcError (Paths.grep, [t, toText f])
+    grp t f = sysN' Paths.grep [t, toText f] id
+    -- grep, but expecting an exit 1
+    grp' t f = sysN' Paths.grep [t, toText f] xx
+--    grp' t f = sysN @ProcError (toMLCmdSpec (Paths.grep, [t, toText f], xx))
+    check name t p =
+      (name, \ f → ѥ (grp t f) ≫ \ case 𝕽 r → p r
+                                        𝕷 e → assertFailure (show e)
+      )
+    check' name t p =
+      (name, \ f → ѥ (grp' t f) ≫ \ case 𝕽 r → p r
+                                         𝕷 e → assertFailure (show e)
+      )
+
+    assertSuffix t x = assertBool ([fmt|'%t' should be a suffix of '%t'|] t x)
+                                  (t `isSuffixOf` x)
+    assertBSSuffix t x = assertBool ([fmt|'%w' should be a suffix of '%w'|] t x)
+                                    (t `BS.isSuffixOf` x)
+  in
+    [ testsWithTempfile foo
+                        [ check' "()" "xxx"
+                                       ((() @=?) ∘ snd)
+                        , check "Text" "mar"
+                                       ((("martyn 12\nmarbyns 3\n"∷𝕋)@=?) ∘ snd)
+                        , check "Bytestring" "mar"
+                                       ((("martyn 12\nmarbyns 3\n"∷𝔹𝕊)@=?) ∘snd)
+                        , check "[Text]" "mar"
+                                (((["martyn 12","marbyns 3"∷𝕋]) @=?) ∘ snd)
+                        ]
+    , -- use binary input here to generate a stderr msg from grep
+      testsWithTempfile ("\x000"⊕foo)
+                        [ check "([Text],[Text])" "mar"
+                                 (\ (_,(o,[e])) → do
+                                     (([]∷[𝕋]) @=? o)
+                                     assertSuffix "binary file matches" e
+                                 )
+                        ]
+    , testsWithTempfile ("\x000"⊕foo)
+                        [ check "([Text],ByteString)" "mar"
+                                 (\ (_,(o,e)) → do
+                                     (([]∷[𝕋]) @=? o)
+                                     assertBSSuffix "binary file matches\n" e
+                                 )
+                        ]
+    ]
+----------------------------------------
 
 grep_ ∷ (MonadIO μ, MLMakeIStream σ,
          MonadError ProcError μ, MonadLog (Log MockIOClass) μ) ⇒
@@ -196,10 +282,12 @@ logMatches r ls =
 
 ----------------------------------------
 
-tests ∷ TestTree
-tests =
+procTests ∷ TestTree
+procTests =
   let
-    p ∷ 𝕋 → IO (𝔼 ProcError ProcResult)
+    foo ∷ 𝕋
+    foo = unlines [ "jimmy 7", "martyn 12", "marbyns 3" ]
+    p   ∷ 𝕋 → IO (𝔼 ProcError ProcResult)
     p t = mkProcResult ⊳⊳ (ѥ @ProcError $ grep t foo)
     logTests ∷ Word8 → [(TestName,[LogEntry MockIOClass] → Assertion)]
     logTests e = [ ([fmt|exit %d|] e, logIncludes ([fmt|Execution exit %d|] e))
@@ -282,6 +370,9 @@ tests =
       ]
 
 ----------------------------------------
+
+tests ∷ TestTree
+tests = testGroup "MonadIO.T.Process" [ procTests, sysTests ]
 
 _test ∷ IO ExitCode
 _test = runTestTree tests